Cast int to unsigned int in BoardManager test.

+++ Testing/Core/BoardManager/BoardManager-Test.cpp
@@ -32,7 +32,7 @@ TEST(BoardManagerTests, gameInit)
             }
             else // expect all other locations to be empty
             {
-                EXPECT_EQ(nullptr, boardGrid[i][j]); 
+                EXPECT_EQ(nullptr, boardGrid[i][j]);
             }
         }
     }
@@ -49,8 +49,8 @@ TEST(BoardManagerTests, getTileStack)
     std::queue<const Tile*> tileQueue1 = tileStack->getQueue((unsigned int) 1);
     std::queue<const Tile*> tileQueue2 = tileStack->getQueue((unsigned int) 2);
 
-    EXPECT_EQ(tileQueue1.size(), NUMBER_OF_PLAYABLE_TILES / 2);
-    EXPECT_EQ(tileQueue2.size(), NUMBER_OF_PLAYABLE_TILES / 2);
+    EXPECT_EQ(tileQueue1.size(), (unsigned int)(NUMBER_OF_PLAYABLE_TILES / 2));
+    EXPECT_EQ(tileQueue2.size(), (unsigned int)(NUMBER_OF_PLAYABLE_TILES / 2));
 
     while(!tileQueue1.empty() || !tileQueue2.empty())
     {
@@ -192,7 +192,7 @@ TEST(BoardManagerTests, makeMove)
     EXPECT_EQ(nullptr, Board::get(coord1));
     EXPECT_EQ(tileStack->front(), front);
     EXPECT_FALSE(tile1.isPlaced());
-    
+
     BoardManager::makeMove(move1, 1);
 
     EXPECT_EQ(&tile1, Board::get(coord1));
